Do not add evil eagle alt kill rules if evil eagle is not in D7

+++ logic/dungeon7.py
@@ -34,8 +34,9 @@ class Dungeon7:
             first_key.items[0].forced_item = KEY7
             
         if options.logic == 'hard' or options.logic == 'glitched' or options.logic == 'hell':
-            boss.connect(pre_boss, OR(MAGIC_ROD, AND(BOMB, BOW))) # magic rod and bomb arrows allow a 3 cycle which avoids the feather wind attack
-            boss.connect(pre_boss, AND(BOW, SHIELD)) # limited arrow amount is rough
+            if world_setup.boss_mapping[6] == 6:
+                boss.connect(pre_boss, OR(MAGIC_ROD, AND(BOMB, BOW))) # magic rod and bomb arrows allow a 3 cycle which avoids the feather wind attack
+                boss.connect(pre_boss, AND(BOW, SHIELD)) # limited arrow amount is rough
             
         if options.logic == 'glitched' or options.logic == 'hell':
             topright_pillar_area.connect(entrance, AND(FEATHER, SWORD)) # superjump in the center to get on raised blocks, superjump in switch room to right side to walk down. center superjump has to be low so sword added
@@ -45,7 +46,8 @@ class Dungeon7:
                 bottomleft_owl.connect(bottomleftF2_area, AND(attack_hookshot, STONE_BEAK7)) # sideways block push to get to the owl statue (attack_hookshot is already implied from bottomleftF2_area)
             final_pillar.connect(bottomleftF2_area, BOMB) # bomb trigger pillar
             pre_boss.connect(final_pillar, FEATHER) # superjump on top of goomba to extend superjump to boss door plateau
-            boss.connect(pre_boss, OR(BOMB, BOW, HOOKSHOT, MAGIC_ROD)) # use bombs while on staircase to time when evil eagle flies by. Or hit the boss straight above staircase, then use ranged (bow/hookshot/magic rod (l2sword??)) straight down to kill evil eagle offscreen
+            if world_setup.boss_mapping[6] == 6:
+                boss.connect(pre_boss, OR(BOMB, BOW, HOOKSHOT, MAGIC_ROD)) # use bombs while on staircase to time when evil eagle flies by. Or hit the boss straight above staircase, then use ranged (bow/hookshot/magic rod (l2sword??)) straight down to kill evil eagle offscreen
             
             
         if options.logic == 'hell':
